// client/src/types/Inventory.ts
// src/types/Inventory.ts
export interface InventoryItem {
  id: string
  name: string
  currentCount: number
  expectedCount: number
}

export interface InventorySection {
  id: string
  ownerId: string
  title: string
  items: InventoryItem[]
}

export interface NewInventorySection {
  title: string
  items: Omit<InventoryItem, "id">[]
}